feat: Swap init for print_notice
This is a change we can make while we're in limited alpha. In the future
we'll want to make this change in a way that is backwards compatible.

With that in mind we've also added an else condition that will raise an
error if we receive an unknown command, but only if debug is enabled.

Co-authored-by: Nate Vick <nate.vick@hint.io>