Get run id and attempt from build env
These values are sometimes available in the build env and as such we
should attempt to use them. If they are not available, we can generate
a default where necessary.

The build env code is due for a refactor, but this is a quick fix to
get the values we need.

Co-authored-by: Nate Vick <nate.vick@hint.io>