Ensure runner id is safe for filenames & implement wait_for_connectivity
We discovered a scenario where things would hang if the transport
crashed before it could open the write pipe. This commit adds a
wait_for_connectivity method that will wait for the transport to
open the pipe for writing. If it doesn't happen within 5 seconds,
the controller will exit.

We also discovered that the runner id could contain characters
that are not allowed in filenames. This commit adds a safe_filename
method that will strip out any characters that are not allowed in
filenames.