Fix two issues relating to a bug that caused runners not to exit
A NoMethodError was being raised when we called `runner.finish` in the
`handle_close` method of the controller. This was preventing the runner
from exiting, and it wasn't clear why because the rescue was swallowing
the error.

Our intent is to rescue no method errors when there is no handler
defined for a provided command. The new approach is to use respond_to
to determine this vs rescuing all NoMethodErrors.

Co-authored-by: Nate Vick <nate.vick@hint.io>