Set a static run_attempt for Semaphore
On Semaphore, hitting the rerun button generates an entirely new build
with a unique job id, workflow id, and pipeline id. As such, we don't
have anything to group it by in order to tell if it is a rerun or not.

Initially I figured I'd live the run_attempt blank since we generate one
if it isn't set. But, then we have the problem were a straggler node can
restart a build if it comes online after the build has finished.

With this in mind, I think it's better to explicitly set a static
run_attempt for Semaphore since each build is a one-shot.

As an aside, perhaps we should consider ditching our auto-generated
after all.

Co-authored-by: Nate Vick <nate.vick@hint.io>