Publish ruby release without compiled transport included
Bundler requires a generic ruby release in the event that the 'ruby'
platform is included in PLATFORMS of a Gemfile.lock. We plan to support
all common platforms, but we can't include a precompiled binary in the
ruby release.

So, we will raise an error telling the user to reach out to support or
compile the binary themselves (advanced).

Co-authored-by: Nate Vick <nate.vick@hint.io>